added card actions, replaced homepage
