+++ package.json
@@ -1,4 +1,4 @@
-{ 
+{
   "type": "module",
   "name": "orbital",
   "version": "0.1.0",
