+++ src/pages/Home.js
@@ -1,73 +1,66 @@
-import { Drawer, Grid, makeStyles, Paper } from '@material-ui/core';
-import React, { useEffect, useState } from 'react';
-import ProjectCard from '../components/ProjectCard';
-import Search from '../components/Search';
-import { db } from '../firebase'
+import React from 'react';
+import algoliasearch from 'algoliasearch';
+import { InstantSearch, connectHits, connectSearchBox} from 'react-instantsearch-dom';
+import PostCard from '../components/PostCard'
+import { Grid, TextField, Button } from '@material-ui/core';
 
-const useStyles = makeStyles((theme) => {
-    return {
-        homeResults: {
-            height: '100%',
-            width: '80%',            
-            margin: 'auto auto',
-        },
-    }
-});
+const searchClient = algoliasearch(
+  'ES79ODFVNM',
+  'c57f19049ad61dad541fc8f7659c0f92'
+);
+
+const Hits = ({ hits }) => (
+    <Grid container spacing={3} style={{width: '80%', margin: 'auto auto'}} >
+      {hits.map(hit => (
+        <Grid item xs={12} sm={6}  key={hit.objectID}>
+            <PostCard hit={hit} />
+        </Grid>
+      ))}
+    </Grid>
+);
+
+const CustomHits = connectHits(Hits);
+
+const SearchBox = ({ isSearchStalled, refine }) => (
+    <Grid item  style={{marginBottom: '20px'}} xs={8} >
+        <form noValidate action="" role="search">
+            <Grid container>
+                <Grid item xs={8}>
+                <TextField 
+                    style={{background: "white", borderRadius: "4px"}}
+                    variant="outlined"
+                    label="Search"
+                    fullWidth
+                    onChange={e => refine(e.currentTarget.value)}
+                />
+                </Grid>
+                <Grid xs={4}>
+                    <Button onClick={() => refine('')} color="primary" variant="contained" style={{marginLeft: '20px', marginTop: '0px', height: '55px', width: '100px', color: 'white'}}>
+                        reset
+                    </Button>
+                </Grid>
+            </Grid>
+            {isSearchStalled ? 'My search is stalled' : ''}
+        </form>
+    </Grid>
+);
+
+const CustomSearchBox = connectSearchBox(SearchBox);
 
 export default function Home() {
-    const classes = useStyles();
-    const [filter, setFilter] = useState(false);
-    const [posts, setPosts] = useState([]);
-    const [render, setRender] = useState(false)
-
-    //for filter pop up
-    const toggleFilter = (open) => (event) => {
-        if (event.type === 'keydown' && (event.key === 'Tab' || event.key === 'Shift')) {
-            return;
-        }      
-        setFilter(open)
-    }
-
-    //sends query to backend when first mounting
-    useEffect(() => {
-        db.collection('posts').onSnapshot((snapShot) => {
-            setPosts(snapShot.docs.map((doc) => ({
-                id: doc.id, data: doc.data()
-            })))
-            setRender(true)
-        })
-    }, [])
 
     return (
-        <div>
-            { !render && <div>Loading..</div> }
-            { render && 
+        <div style={{marginTop: '50px'}}>
+             <InstantSearch
+                indexName="posts"
+                searchClient={searchClient}
+            > 
+            <Grid container spacing={2} justify="center" style={{marginBottom: "20px"}}>
+                <CustomSearchBox />
+            </Grid>
 
-            <div className={classes.homeResults}>
-                <Search toggleFilter={toggleFilter} />
-                <Drawer anchor="right" open={filter} onClose={toggleFilter(false)}>
-                    <Paper style={{height: '1000px', width: '500px', padding: '100px'}} />
-                </Drawer>  
-                <Grid container spacing={3}>
-                    {posts.map(({id, data}) => {
-                        return ( 
-                        <Grid item xs={12} md={6}key={id}>
-                            <ProjectCard
-                                key={id}
-                                id={id}
-                                title={data.title}
-                                author={data.name}
-                                authorId={data.author}
-                                description={data.description}
-                                chips={data.skills}
-                                bookmarkedBy={data.bookmarkedBy}
-                            />
-                        </Grid>
-                    )})}
-                </Grid>
-            </div>
-            }
+            <CustomHits />
+            </InstantSearch>
         </div>
-    );
-  }
-  
+    )
+}
